Fixing validations for the Reason field, and updating some CSS for customchatcard
Added reason field to validations: Updated ContentAttributeValidator to include the new "reason" field in the allowed keys for custom card items.

Added custom_fields field to validations: Also included the custom_fields property in the allowed keys, which was missing and might have been causing errors if you were trying to use that.