feat: Auto-open pending conversations on agent/team assignment
The changes have been applied to app/models/concerns/assignment_handler.rb.
Now, when a conversation that is pending gets assigned to an agent (i.e., assignee_id is set to a present value) or a team (i.e., team_id is set to a present value), the bot_handoff! method will be called. This will, in turn, call open!, changing the conversation's status to open and dispatching the CONVERSATION_BOT_HANDOFF event.
This should address the issue where conversations remain pending after being assigned to an agent via your n8n workflow.